Add macros for ad-hoc tracing while debugging.
Often, an execution flow is needed while debugging specific problems.
The easiest, quickest, and often most reliable way is to add log
messages showing the flow. As opposed to single-step debugging (which
is about the worst way to do any kind of debugging), this way doesn't
mess up timing too much and provides a documented flow that can be
stored away for later comparison.

In the past, simple msg_info() calls were added to achieve this,
involving lots of copying of these calls all over the place, and
sometimes mixing them with new, intended msg_info() calls. The trace
messages are almost always completely removed once the inspected
problem has been resolved, which is a bit tedious in case other
msg_info() calls are present as well. The new macros make it easy to
add and remove traces to/from code.

Change-Id: Id2bdf843098df1c2922bafa5874576d3bbd44737