messages: Add compile-time option for disabling trace messages.
This is useful when debugging. If trace messages had to be added for
runtime analyses, some unit tests are likely to fail because of the
unexpected messages. Disabling the trace messages all at once allows
switching from debugging with runtime message over to unit tests.

Note that code which calls MSG_TRACE() or MSG_TRACE_FORMAT() is
*never* supposed to be checked into the repository; therefore, the new
MSG_TRACE_ENABLED option can always be enabled. Trace messages are
always of temporary nature and are solely meant for aiding debugging
sessions.

Change-Id: Icc19f6f0bd8c820c82be96844bc93715151294e3