messages: Add macros for various standard situations.
New macros:

  * BUG_IF(): Like BUG(), but with a condition. This makes bug
    messages more concise.

  * TODO(): Like BUG(), but prefixed with "TODO:".

  * MSG_UNREACHABLE(): Bug message telling that some code was reached
    which, by design, shouldn't have been reached (including function
    name and line number).

  * MSG_NOT_IMPLEMENTED(): Like MSG_UNREACHABLE(), but as a TODO
    message and telling that something hasn't been implemented yet.

  * MSG_TRACE_THIS(): Like MSG_TRACE(), but automatically dumps the
    value of the "this" pointer along with the trace output.
    Obviously, the applicability of this macro is restricted to C++
    function members.

The macros which report code locations do not report file names
because the names that __FILE__ expands to tend to be very long when
built out of tree (and especially so with Yocto Project). Refusing to
put in the names manually, they are simply not reported at all.
Function names should be good enough for this purpose.

Change-Id: I4c418c00a1214bf6e97c6ec29956c8eadd180bfc