Unit tests: Check expectations in defined order.
This works for all mocks together now, not only for expectations of
individual mocks. We have new MockExpectationSequence class which
records the sequence of expectations as they are added in the tests.
Expectation objects have a serial number now, and when running the
tests, the expectation object serials are checked whether or not they
form a proper sequence.

To minimize impact on existing unit tests, it is possible to use a
singleton MockExpectationSequence object which is used by default by
all mocks which support this. Unit test code needs to define the
symbol MOCK_EXPECTATION_WITH_EXPECTATION_SEQUENCE_SINGLETON to enable
the feature, and it must declare the singleton object. It must be
initialized before any other mock objects are created (because their
constructors refer to the singleton object), and should be reset()
before each test is started. Expectations should be checked after at
test has run via done() member function, before checking the other
mocks.

Change-Id: Ib2c8f92cd71821c0c205b6f8c942b11eafde5dbb