messages: MSG_ or msg_ prefixes for all macros/functions.
This commit obsoletes the macros BUG(), BUG_IF(), TODO(),
APPLIANCE_BUG(), and log_assert(). Their new names are MSG_BUG(),
MSG_BUG_IF(), etc. The primary driver behind this change is Doctest,
which declares a function member named "log_assert", and this fails to
compile in case log_assert is a macro. Instead of introducing yet
another #ifdef hell in order to work around this, we simply clean up
and use a proper prefix.

For a quick cure in older projects, it is possible to revive the
legacy macro names by defining MSG_ENABLE_LEGACY_MACROS. We will
remove this option as soon as we can manage.

Change-Id: I860a9e9654c0b0d092e125c0f9599120c32ff075