taddybus: Allow use of synchronous D-Bus calls.
Synchronous calls are processed directly, bypassing the D-Bus main
loop and taking a code path which is entirely different from the
asynchronous case.

This is useful for calls which are *known* to return fast, and is safe
for calls which are *guaranteed* not to call back the calling process,
neither directly nor indirectly. Lack of such guarantees may lead to
deadlocks spanning several process, which are notoriously hard to
debug.

Avoid synchronous calls where it is possible!

Change-Id: I820f831357356914804eb9ec660e39794076c018