mock_expectation: Support rearrangement of expectation sequences.
If, for some reason, it is hard it define expectations in the correct
order (maybe do to convenience functions which define expectations,
but do not allow to insert extra expectations on certain conditions),
the unit test author may now reorder the expectations added to a
MockExpectationSequence.

To this end, member functions move_range(), swap_slots(), and
reverse_last_slots() have been added. The functions make sure that
they only work on expectations that have not been checked yet.

Change-Id: If5557b6e85ef2a3c72d2e49df3b708ba1374331e