mock: Refactor to enable easy future extensions.
The MockExpectationsTemplate which was formerly embedded in the
various specialized mock classes has been turned into MockBase, a
class which mock classes may derive from. That is, mocks are not
composed with MockExpectationsTemplate, but derived from MockBase,
which, in this case, is the better choice.

In addition, MockExpectationBase was added which all specialized
expectation classes may derive from. This also makes
MockExpectationSequence a little bit simpler.

These changes enable fixing a bug when rearranging expectation
sequences which involve the same mock. The registered expectations
must be sorted by sequence number within the mock after having them
shuffled around in the MockExpectationSequence, otherwise they will
appear in the wrong order when checking expectations.

Another extension is the generation of an expectation string in case
the unit test runs into an unexpected function call for which an
expectation would be required. It is now easy to identify which
expectation is missing and how it may look like based on the
parameters passed by the code under test.

Change-Id: I8798d04f91b5ef30b5688c44f3e8ac5c25918a1b